refactor(upload): dedupe image resize/TIFF pipeline, delete dead ImageUploader
ImageUploader.jsx had zero live consumers (only a barrel export) but
duplicated ~150 lines of TIFF-handling and canvas resize/JPEG re-encode
logic that also lived, near-verbatim, in UnifiedUploader.jsx. A third
hand-copied version of the resize/re-encode step lived in the Office
add-in's buildChatApiMessages.js and had already drifted (stale comment
referencing the now-deleted component).

- Move the shared logic into processImageFile()/resizeImageCanvas() in
  features/upload/utils/fileProcessing.js.
- UnifiedUploader now delegates to processImageFile instead of a local
  copy; output shape is unchanged.
- buildChatApiMessages's resizeImageBase64 now calls the shared
  resizeImageCanvas primitive instead of hand-rolling the same canvas
  block.
- Delete ImageUploader.jsx and its barrel export.

Fixes #1755
