--- client/src/features/upload/components/ImageUploader.jsx
@@ -1,268 +1,0 @@
-import { useTranslation } from 'react-i18next';
-import Icon from '../../../shared/components/Icon';
-import Uploader from './Uploader';
-import './ImageUpload.css';
-import { processTiffFile } from '../utils/fileProcessing';
-
-/**
- * Lightweight wrapper for uploading images.
- */
-export default function ImageUploader({
-  onImageSelect,
-  disabled = false,
-  imageData = null,
-  config = {}
-}) {
-  const { t } = useTranslation();
-  const MAX_FILE_SIZE_MB = config.maxFileSizeMB || 10;
-  const SUPPORTED_FORMATS = config.supportedFormats || [
-    'image/jpeg',
-    'image/jpg',
-    'image/png',
-    'image/gif',
-    'image/webp'
-  ];
-  const RESIZE_IMAGES = config.resizeImages !== false;
-  const MAX_DIMENSION = config.maxResizeDimension || 1024;
-
-  const processImage = async file => {
-    // Check if this is a TIFF file
-    const isTiff = file.type === 'image/tiff' || file.type === 'image/tif';
-
-    if (isTiff) {
-      try {
-        // Process TIFF file and convert to PNG
-        const pages = await processTiffFile(file, {
-          maxDimension: MAX_DIMENSION,
-          resize: RESIZE_IMAGES
-        });
-
-        // For multipage TIFFs, return all pages as separate images if supported
-        // Note: ImageUploader doesn't support allowMultiple, so we'll use first page only
-        // but still return multipleResults for future compatibility
-        if (pages.length > 1) {
-          // Prepare all pages
-          const pageResults = [];
-
-          for (let i = 0; i < pages.length; i++) {
-            const page = pages[i];
-
-            // Create blob URL for preview
-            const response = await fetch(page.base64);
-            const blob = await response.blob();
-            const previewUrl = URL.createObjectURL(blob);
-
-            // Generate filename with page number
-            const baseFileName = file.name.replace(/\.tiff?$/i, '');
-            const fileName = `${baseFileName}_page${page.pageNumber}.png`;
-
-            pageResults.push({
-              preview: previewUrl,
-              data: {
-                base64: page.base64,
-                fileName: fileName,
-                fileSize: blob.size,
-                fileType: 'image/png',
-                width: page.width,
-                height: page.height,
-                originalFileType: file.type,
-                originalFileName: file.name,
-                pageNumber: page.pageNumber,
-                totalPages: page.totalPages
-              }
-            });
-          }
-
-          // Return first page as primary, but include all pages info
-          return {
-            preview: pageResults[0].preview,
-            data: {
-              ...pageResults[0].data,
-              tiffPages: pages
-            }
-          };
-        }
-
-        // For single-page TIFF
-        const firstPage = pages[0];
-
-        // Create blob URL for preview
-        const response = await fetch(firstPage.base64);
-        const blob = await response.blob();
-        const previewUrl = URL.createObjectURL(blob);
-
-        return {
-          preview: previewUrl,
-          data: {
-            base64: firstPage.base64,
-            fileName: file.name.replace(/\.tiff?$/i, '.png'),
-            fileSize: blob.size,
-            fileType: 'image/png',
-            width: firstPage.width,
-            height: firstPage.height,
-            originalFileType: file.type,
-            originalFileName: file.name
-          }
-        };
-      } catch (error) {
-        console.error('Error processing TIFF file:', error);
-        throw new Error('tiff-processing-error');
-      }
-    }
-
-    return new Promise((resolve, reject) => {
-      const reader = new FileReader();
-
-      reader.onload = e => {
-        const img = new Image();
-        img.onload = () => {
-          const previewUrl = URL.createObjectURL(file);
-
-          // If resizing is disabled, simply return the original image data
-          if (!RESIZE_IMAGES) {
-            return resolve({
-              preview: previewUrl,
-              data: {
-                base64: e.target.result,
-                fileName: file.name,
-                fileSize: file.size,
-                fileType: file.type,
-                width: img.width,
-                height: img.height
-              }
-            });
-          }
-
-          const canvas = document.createElement('canvas');
-          let width = img.width;
-          let height = img.height;
-
-          if (width > height && width > MAX_DIMENSION) {
-            height = Math.round((height * MAX_DIMENSION) / width);
-            width = MAX_DIMENSION;
-          } else if (height > MAX_DIMENSION) {
-            width = Math.round((width * MAX_DIMENSION) / height);
-            height = MAX_DIMENSION;
-          }
-
-          canvas.width = width;
-          canvas.height = height;
-          const ctx = canvas.getContext('2d');
-          ctx.drawImage(img, 0, 0, width, height);
-          const base64 = canvas.toDataURL('image/jpeg', 0.8);
-
-          resolve({
-            preview: previewUrl,
-            data: {
-              base64,
-              fileName: file.name,
-              fileSize: file.size,
-              fileType: 'image/jpeg',
-              width,
-              height
-            }
-          });
-        };
-        img.onerror = () => reject(new Error('invalid-image'));
-        img.src = e.target.result;
-      };
-
-      reader.onerror = () => reject(new Error('read-error'));
-      reader.readAsDataURL(file);
-    });
-  };
-
-  const getErrorMessage = code => {
-    switch (code) {
-      case 'file-too-large':
-        return t('errors.fileTooLarge', {
-          maxSize: MAX_FILE_SIZE_MB,
-          defaultValue: `File too large. Maximum size is ${MAX_FILE_SIZE_MB}MB.`
-        });
-      case 'unsupported-format':
-        return t('errors.unsupportedFormat', {
-          formats: SUPPORTED_FORMATS.map(f => f.replace('image/', '.')).join(', '),
-          defaultValue: `Unsupported file format. Please use: ${SUPPORTED_FORMATS.map(f => f.replace('image/', '.')).join(', ')}`
-        });
-      case 'invalid-image':
-        return t('errors.invalidImage', 'Invalid image file');
-      case 'tiff-processing-error':
-        return t(
-          'errors.tiffProcessingError',
-          'Error processing TIFF file. The file may be corrupted or in an unsupported TIFF format.'
-        );
-      case 'read-error':
-        return t('errors.readError', 'Error reading file');
-      default:
-        return t('errors.fileProcessingError', 'Error processing file. Please try again.');
-    }
-  };
-
-  return (
-    <Uploader
-      accept={SUPPORTED_FORMATS}
-      maxSizeMB={MAX_FILE_SIZE_MB}
-      disabled={disabled}
-      onSelect={onImageSelect}
-      onProcessFile={processImage}
-      data={imageData}
-    >
-      {({ preview, error, isProcessing, handleButtonClick, handleClear, inputProps }) => (
-        <div className="image-uploader">
-          <input {...inputProps} />
-          {preview ? (
-            <div className="relative mt-2 mb-4">
-              <div className="relative rounded-lg overflow-hidden border border-gray-300">
-                <img
-                  src={preview}
-                  alt={t('common.preview', 'Preview')}
-                  className="max-w-full max-h-60 mx-auto"
-                />
-                <button
-                  type="button"
-                  onClick={handleClear}
-                  className="absolute top-2 right-2 bg-gray-800 bg-opacity-70 text-white rounded-full p-1 hover:bg-opacity-90"
-                  title={t('common.remove', 'Remove image')}
-                  aria-label={t('common.remove', 'Remove image')}
-                >
-                  <Icon name="x" className="w-4 h-4" aria-hidden="true" />
-                </button>
-              </div>
-              <div className="text-xs text-gray-500 mt-1 text-center">
-                {t('components.imageUploader.imageSelected', 'Image selected')}
-              </div>
-            </div>
-          ) : (
-            <div className="mt-2 mb-4">
-              <button
-                type="button"
-                onClick={handleButtonClick}
-                disabled={disabled || isProcessing}
-                className={`flex items-center justify-center w-full p-3 border-2 border-dashed rounded-lg ${
-                  disabled || isProcessing
-                    ? 'border-gray-300 dark:border-gray-600 bg-gray-100 dark:bg-gray-800 text-gray-400 dark:text-gray-500 cursor-not-allowed'
-                    : 'border-gray-400 dark:border-gray-500 hover:border-indigo-500 dark:hover:border-indigo-400 hover:text-indigo-500 dark:hover:text-indigo-400 bg-white dark:bg-gray-800'
-                }`}
-              >
-                <Icon name="camera" className="mr-2" />
-                <span>{t('components.imageUploader.uploadImage', 'Upload Image')}</span>
-              </button>
-
-              {error && <div className="text-red-500 text-sm mt-1">{getErrorMessage(error)}</div>}
-
-              <div className="text-xs text-gray-500 dark:text-gray-400 mt-1 text-center">
-                {t(
-                  'components.imageUploader.supportedFormats',
-                  'Supported: JPG, PNG, GIF, WebP (max {{maxSize}}MB)',
-                  {
-                    maxSize: MAX_FILE_SIZE_MB
-                  }
-                )}
-              </div>
-            </div>
-          )}
-        </div>
-      )}
-    </Uploader>
-  );
-}
