+++ client/src/features/upload/components/index.js
@@ -1,5 +1,4 @@
 export { default as AttachedFilesList } from './AttachedFilesList';
 export { default as CloudStoragePicker } from './CloudStoragePicker';
-export { default as ImageUploader } from './ImageUploader';
 export { default as UnifiedUploader } from './UnifiedUploader';
 export { default as Uploader } from './Uploader';
